% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_not.R
\name{conflict_not}
\alias{conflict_not}
\title{Indicates where a function is loaded from}
\usage{
conflict_not()
}
\description{
Call this function as an addin over selected text to add packagename::
before any function call.
}
