% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_make.R
\name{data_make}
\alias{data_make}
\title{Easy generation of data with dependent and independent variables}
\usage{
data_make(n_dependent = 1, n_random = 1, n = 100)
}
\arguments{
\item{n_dependent}{Integer. Number of variables that are correlated with
the response.}

\item{n_random}{Integer. Number of variables that isn't correlated with
the response.}

\item{n}{Integer. Number of observations.}
}
\value{
data.frame with 1 + n_dependent + n_random number of variables.
}
\description{
This function is meant ease testing of other functions by creating a
dataframe with a variable number of correlated variables and independent
variables.
}
\examples{
data_make()

data_make(2, 5, 1000)
}
