% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehlib-package.R
\docType{package}
\name{ehlib-package}
\alias{ehlib}
\alias{ehlib-package}
\title{ehlib: Personal R pacakge}
\description{
Package of personal commonly used R functions.
}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com}

}
\keyword{internal}
