% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_hits.R
\name{google_hits}
\alias{google_hits}
\title{Number of hits of results from google searches}
\usage{
google_hits(string)
}
\arguments{
\item{string}{vector of strings to search for.}
}
\value{
A tibble with two columns. search and hits.
}
\description{
Provided with a (vector of) string(s) will return the number of results
approximately given by google.
}
\examples{
\dontrun{
google_hits(string = letters) \%>\%
  ggplot(aes(search, hits)) +
  geom_col()
}
}
