% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_df.R
\name{kmeans_df}
\alias{kmeans_df}
\title{kmeans function that takes a data.frame and a formula as inputs}
\usage{
kmeans_df(data, formula, ...)
}
\arguments{
\item{data}{A data.frame.}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
can be coerced to that class)}

\item{\dots}{Additional arguments to be passed to the kmeans function.}
}
\value{
an object of class "kmeans".
}
\description{
This function is a simple wrapper that lets you work with \code{kmeans()}
while using data.frame, thus allowing you to stay with the same format.
}
\examples{
kmeans_df(~., cars, centers = 2)
}
