% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_df.R
\name{knn_df}
\alias{knn_df}
\title{knns function from class that takes a data.frame and a formula as inputs}
\usage{
knn_df(data, formula, split, ...)
}
\arguments{
\item{data}{A data.frame.}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
can be coerced to that class)}

\item{split}{variable that was used to split data.frame into training and
test set.}

\item{\dots}{Additional arguments to be passed to the
\code{\link[class]{knn}} function.}
}
\value{
an object of class "knn".
}
\description{
This function is a simple wrapper that lets you work with \code{class::knns()}
while using data.frame, thus allowing you to stay with the same format.
}
\details{
It allows for the use of the formula expression from base stats, where the
lhs is the factor of the true. The split variable must be logical.
}
\examples{
data <- iris
data$split <- as.logical(sample(0:1, nrow(data), TRUE))

knn_df(data, Species ~ Sepal.Length + Sepal.Width, split, k = 5, prob = TRUE)
}
