% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_scrapers.R
\name{palette_color_hex}
\alias{palette_color_hex}
\title{Scrapes color palette from color-hex.com}
\usage{
palette_color_hex(url)
}
\arguments{
\item{url}{Url of website as string.}
}
\value{
A vector of color values as hex color codes.
}
\description{
Scrapes color palette from color-hex.com
}
\examples{
palette_color_hex("http://www.color-hex.com/color-palette/185")
}
