% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_scrapers.R
\name{palette_coolors}
\alias{palette_coolors}
\title{Scrapes color palette from coolors.co}
\usage{
palette_coolors(url)
}
\arguments{
\item{url}{Url of website as string.}
}
\value{
A vector of color values as hex color codes.
}
\description{
Scrapes color palette from coolors.co
}
\examples{
palette_coolors("https://coolors.co/app/05668d-028090-00a896-02c39a-f0f3bd")
}
