% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztpois.R
\name{rztpois}
\alias{rztpois}
\title{The Zero-truncated Poisson distribution}
\usage{
rztpois(n, lambda = 1)
}
\arguments{
\item{n}{number of random values to return.}

\item{lambda}{vector of (non-negative) means.}
}
\value{
Generates random deviates of the zero truncated Poisson.
}
\description{
Random generation for the Poisson distribution with parameter lambda.
}
\examples{
rztpois(100)
rztpois(1000, 5)
}
