% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_rep.R
\name{scale_rep}
\alias{scale_fill_rep}
\alias{scale_alpha_rep}
\alias{scale_color_rep}
\alias{scale_colour_rep}
\alias{scale_linetype_rep}
\alias{scale_shape_rep}
\alias{scale_size_rep}
\title{Create your own discrete scale}
\usage{
scale_fill_rep(..., values)

scale_alpha_rep(..., values)

scale_color_rep(..., values)

scale_colour_rep(..., values)

scale_linetype_rep(..., values)

scale_shape_rep(..., values)

scale_size_rep(..., values)
}
\arguments{
\item{...}{Other arguments passed on to [discrete_scale()] to control name, limits,
breaks, labels and so forth.}

\item{values}{a set of aesthetic values to map data values to. If this
is a named vector, then the values will be matched based on the names.
If unnamed, values will be matched in order (usually alphabetical) with
the limits of the scale. Will repeat if number of labels is less.}
}
\description{
This allows you to specify you own set of mappings from levels in the
data to aesthetic values that will repeat if a insufficient number of levels
 is provided.
}
\examples{
library(ggplot2)

dat <- data.frame(x = sample(LETTERS, 200, TRUE))

ggplot(dat, aes(x = x)) +
  geom_bar(aes(fill = x)) +
    scale_fill_rep(values = rainbow(5))
}
