% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_after}
\alias{str_after}
\title{Extract text after pattern (not including)}
\usage{
str_after(string, pattern)
}
\arguments{
\item{string}{A number. Input vector.}

\item{pattern}{Pattern to match after.}
}
\value{
A character vector.
}
\description{
Extract text after pattern (not including)
}
