% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_before}
\alias{str_before}
\title{Extract text before pattern (not including)}
\usage{
str_before(string, pattern)
}
\arguments{
\item{string}{A character vector. Input vector.}

\item{pattern}{Pattern to match before.}
}
\value{
A character vector.
}
\description{
Extract text before pattern (not including)
}
\examples{
string <- "This is a test string where we do the testing."

str_before(string, "where")
str_before(string, "test")
}
