% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_before_first}
\alias{str_before_first}
\title{Before first}
\usage{
str_before_first(string, pattern)
}
\arguments{
\item{string}{A character vector. Input vector.}

\item{pattern}{Pattern to look after.}
}
\value{
A character vector.
}
\description{
Before first
}
\examples{
string <- "This is a test string where we do the testing."

str_before_first(string, "where")
str_before_first(string, "test")
}
