% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_between_n}
\alias{str_between_n}
\title{Extract nth text between two patterns (not including)}
\usage{
str_between_n(string, start, end, n)
}
\arguments{
\item{string}{A character vector. Input vector.}

\item{start}{Pattern to match after.}

\item{end}{Pattern to match before.}

\item{n}{Integer.}
}
\value{
A character vector.
}
\description{
Extract nth text between two patterns (not including)
}
