% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_nth_word}
\alias{str_nth_word}
\title{Extract the nth word}
\usage{
str_nth_word(string, n, sep = " ")
}
\arguments{
\item{string}{A character vector. Input vector.}

\item{n}{Integer. Which word to extract.}

\item{sep}{Pattern to seperate words with.}
}
\value{
A character vector.
}
\description{
Fairly rough function. Ideal to work with n-grams.
}
