% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_asm.R
\name{tidy_masm}
\alias{tidy_masm}
\title{Tidy output of the gvlma function}
\usage{
tidy_masm(x, model)
}
\arguments{
\item{x}{lm object or data.frame with lm objects as a list-column.}

\item{model}{Column name of the}
}
\value{
data.frame with 18 columns.
}
\description{
This function is meant takes a lm object, or a data.frame of lm objects and
returns a data.frame of lm model assumption tests outlined in “Global
validation of linear model assumptions”.
}
\examples{
\dontrun{
data <- car::Prestige

data \%>\%
  nest() \%>\%
  mutate(lm = map(data, ~ lm(education ~ income, .x))) \%>\%
  tidy_masm(lm)

  lm(education ~ income, data) \%>\%
    tidy_masm()
}
}
