% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{Transform a catagorical variable to one hot representation}
\usage{
one_hot(data, var)
}
\arguments{
\item{data}{a data.frame or tibble.}

\item{var}{name of catagorical variable.}
}
\value{
A data.frame or tibble with the catagorical variable replaced with
 an one hot representation.
}
\description{
The one hot representation is analogous to factors in the sense that each
value of the catagorical variable have its own logical vector that is 1
when present and 0 when it isn't.
}
\examples{
one_hot(iris, Species)
}
