% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_new.R
\name{step_new}
\alias{step_new}
\title{Re-imagination of stats::step() function}
\usage{
step_new(object, steps = 1000, trace = TRUE)
}
\arguments{
\item{object}{an object representing a model of an appropriate class
(mainly "lm" and "glm"). This is used as the initial model in the
stepwise search.}

\item{steps}{the maximum number of steps to be considered. The default is
1000 (essentially as many as required). It is typically used to stop the
process early.}

\item{trace}{Logical. If TRUE prints information of steps taking. Defaults
to TRUE.}
}
\value{
to be determined
}
\description{
This functions aims to add the same functionality as stats::step() with the
adition to saving intermediate models along with diagnogists.
}
\examples{
dataset <- data_make()

object <- lm(response ~ ., data = dataset)

X <- step_new(object)

dataset <- data_make(n_dependent = 5, n_random = 5, n = 1000)

object <- lm(response ~ ., data = dataset)

X <- step_new(object)
}
