% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_extensions.R
\name{str_after_last}
\alias{str_after_last}
\title{Extract after last}
\usage{
str_after_last(string, pattern)
}
\arguments{
\item{string}{A character vector. Input vector.}

\item{pattern}{Pattern to look after.}
}
\value{
A character vector.
}
\description{
Extract after last
}
